test(ui): move the list, notices, quick and poller tests beside their files
fulllayout_test.go had grown into a second home for tests that cover
listview.go, notices.go, quick.go and poller.go. Each one moves to the
file named for what it calls, leaving the layout toggle, focus framing
and pane sizing tests together, and the control-capture test joins
matchExecShape in focuswatch_test.go.
